requirements.txt
clip
torch
torchvision
numpy
pillow